% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkg-package.R
\docType{package}
\name{mypkg-package}
\alias{mypkg}
\alias{mypkg-package}
\title{mypkg: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
